Expose GetObject to API users
This was hidden before because it may result in a low-level detail
being exposed, but upon further thought, this is a sane and reasonable
function to expose.

To add API symmetry, a SaveObject function was added to write data back
to the token. The only method calling the underlying ykpiv function,
`ykpiv_save_object`, Slot.Update, has been refactored to use SaveObject.

Callers should be careful to not write things they shouldn't write in
places they shouldn't write things. Advice for life, really.